Move retention to separate file
