+++ tempodb/tempodb_test.go
@@ -122,60 +122,6 @@ func TestNilOnUnknownTenantID(t *testing.T) {
 	assert.Nil(t, err)
 }
 
-func TestRetention(t *testing.T) {
-	tempDir, err := ioutil.TempDir("/tmp", "")
-	defer os.RemoveAll(tempDir)
-	assert.NoError(t, err, "unexpected error creating temp dir")
-
-	r, w, c, err := New(&Config{
-		Backend: "local",
-		Local: &local.Config{
-			Path: path.Join(tempDir, "traces"),
-		},
-		WAL: &wal.Config{
-			Filepath:        path.Join(tempDir, "wal"),
-			IndexDownsample: 17,
-			BloomFP:         .01,
-		},
-		BlocklistPoll: 0,
-	}, log.NewNopLogger())
-	assert.NoError(t, err)
-
-	c.EnableCompaction(&CompactorConfig{
-		ChunkSizeBytes:          10,
-		MaxCompactionRange:      time.Hour,
-		BlockRetention:          0,
-		CompactedBlockRetention: 0,
-	}, &mockSharder{})
-
-	blockID := uuid.New()
-
-	wal := w.WAL()
-	assert.NoError(t, err)
-
-	head, err := wal.NewBlock(blockID, testTenantID)
-	assert.NoError(t, err)
-
-	complete, err := head.Complete(wal, &mockSharder{})
-	assert.NoError(t, err)
-	blockID = complete.BlockMeta().BlockID
-
-	err = w.WriteBlock(context.Background(), complete)
-	assert.NoError(t, err)
-
-	rw := r.(*readerWriter)
-	// poll
-	checkBlocklists(t, blockID, 1, 0, rw)
-
-	// retention should mark it compacted
-	r.(*readerWriter).doRetention()
-	checkBlocklists(t, blockID, 0, 1, rw)
-
-	// retention again should clear it
-	r.(*readerWriter).doRetention()
-	checkBlocklists(t, blockID, 0, 0, rw)
-}
-
 func TestBlockCleanup(t *testing.T) {
 	tempDir, err := ioutil.TempDir("/tmp", "")
 	defer os.RemoveAll(tempDir)
